chore: Update to using 'setupATLAS -3'
Currently (in March 2023) if setupATLAS is run a warning is printed of

centos7: setupATLAS is python3 environment by default (same as setupATLAS -3).
         If you need the previous python2 environment, do setupATLAS -2.
 	 We strongly encourage API users to migrate their scripts to python3.
          * * * This environment has been setup as python3. * * *

As

```
$ command -v setupATLAS
alias setupATLAS='source ${ATLAS_LOCAL_ROOT_BASE}/user/atlasLocalSetup.sh -3'
```

shows that it isn't required to append the '-3' option to setupATLAS it
is worthwhile to add for now to help signal the Python 3 default.